+++ src/properties.rs
@@ -252,7 +252,7 @@ where
     }
 
     /// Change the display brightness.
-    pub fn change_brightness(&mut self, brightness: Brightness) -> Result<(), DisplayError> {
+    pub fn set_brightness(&mut self, brightness: Brightness) -> Result<(), DisplayError> {
         // Should be moved to Brightness::new once conditions can be used in const functions
         debug_assert!(
             0 < brightness.precharge && brightness.precharge <= 15,